Feed related access control checks

+++ services/feed-manager-service/feed-manager-controller/src/main/java/com/thinkbiganalytics/feedmgr/service/category/DefaultFeedManagerCategoryService.java
@@ -3,12 +3,14 @@ package com.thinkbiganalytics.feedmgr.service.category;
 import com.thinkbiganalytics.feedmgr.InvalidOperationException;
 import com.thinkbiganalytics.feedmgr.rest.model.FeedCategory;
 import com.thinkbiganalytics.feedmgr.rest.model.UserField;
+import com.thinkbiganalytics.feedmgr.security.FeedsAccessControl;
 import com.thinkbiganalytics.feedmgr.service.UserPropertyTransform;
 import com.thinkbiganalytics.metadata.api.Command;
 import com.thinkbiganalytics.metadata.api.MetadataAccess;
 import com.thinkbiganalytics.metadata.api.extension.UserFieldDescriptor;
 import com.thinkbiganalytics.metadata.api.feedmgr.category.FeedManagerCategory;
 import com.thinkbiganalytics.metadata.api.feedmgr.category.FeedManagerCategoryProvider;
+import com.thinkbiganalytics.security.AccessController;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -32,9 +34,14 @@ public class DefaultFeedManagerCategoryService implements FeedManagerCategorySer
     @Inject
     MetadataAccess metadataAccess;
 
+    @Inject
+    private AccessController accessController;
+
     @Override
     public Collection<FeedCategory> getCategories() {
         return metadataAccess.read((Command<Collection<FeedCategory>>) () -> {
+            this.accessController.checkPermission(AccessController.SERVICES, FeedsAccessControl.ACCESS_CATEGORIES);
+
             List<FeedManagerCategory> domainCategories = categoryProvider.findAll();
             return categoryModelTransform.domainToFeedCategory(domainCategories);
         });
@@ -43,6 +50,8 @@ public class DefaultFeedManagerCategoryService implements FeedManagerCategorySer
     @Override
     public FeedCategory getCategoryById(final String id) {
         return metadataAccess.read(() -> {
+            this.accessController.checkPermission(AccessController.SERVICES, FeedsAccessControl.ACCESS_CATEGORIES);
+
             final FeedManagerCategory.ID domainId = categoryProvider.resolveId(id);
             final FeedManagerCategory domainCategory = categoryProvider.findById(domainId);
             return categoryModelTransform.domainToFeedCategory(domainCategory);
@@ -52,6 +61,8 @@ public class DefaultFeedManagerCategoryService implements FeedManagerCategorySer
     @Override
     public FeedCategory getCategoryBySystemName(final String name) {
         return metadataAccess.read(() -> {
+            this.accessController.checkPermission(AccessController.SERVICES, FeedsAccessControl.ACCESS_CATEGORIES);
+
             final FeedManagerCategory domainCategory = categoryProvider.findBySystemName(name);
             return categoryModelTransform.domainToFeedCategory(domainCategory);
         });
@@ -60,6 +71,8 @@ public class DefaultFeedManagerCategoryService implements FeedManagerCategorySer
     @Override
     public void saveCategory(final FeedCategory category) {
         final FeedManagerCategory.ID domainId = metadataAccess.commit(() -> {
+            this.accessController.checkPermission(AccessController.SERVICES, FeedsAccessControl.CREATE_CATEGORIES);
+
             // Determine the system name
             if (category.getId() == null) {
                 category.generateSystemName();
@@ -89,6 +102,8 @@ public class DefaultFeedManagerCategoryService implements FeedManagerCategorySer
 
     @Override
     public boolean deleteCategory(final String categoryId) throws InvalidOperationException {
+        this.accessController.checkPermission(AccessController.SERVICES, FeedsAccessControl.DELETE_CATEGORIES);
+
         final FeedManagerCategory.ID domainId = metadataAccess.read(() -> categoryProvider.resolveId(categoryId));
         categoryProvider.deleteById(domainId);
         return true;
@@ -97,11 +112,17 @@ public class DefaultFeedManagerCategoryService implements FeedManagerCategorySer
     @Nonnull
     @Override
     public Set<UserField> getUserFields() {
-        return metadataAccess.read(() -> UserPropertyTransform.toUserFields(categoryProvider.getUserFields()));
+        return metadataAccess.read(() -> { 
+            this.accessController.checkPermission(AccessController.SERVICES, FeedsAccessControl.ACCESS_CATEGORIES);
+
+            return UserPropertyTransform.toUserFields(categoryProvider.getUserFields()); 
+        });
     }
 
     @Override
     public void setUserFields(@Nonnull Set<UserField> userFields) {
+        this.accessController.checkPermission(AccessController.SERVICES, FeedsAccessControl.CREATE_CATEGORIES);
+
         categoryProvider.setUserFields(UserPropertyTransform.toUserFieldDescriptors(userFields));
     }
 }
